feat: completion, style: PEP8

+++ application.py
@@ -1,23 +1,22 @@
-from flask import Flask, render_template, request, redirect, url_for, jsonify
-from sqlalchemy import create_engine
-from sqlalchemy.orm import sessionmaker
-from database_setup import Base, Category, Item
 
-from flask import session as login_session
-import random, string
-
-from oauth2client.client import flow_from_clientsecrets
-from oauth2client.client import FlowExchangeError
+import random
+import string
 import httplib2
 import json
-from flask import make_response
 import requests
+from flask import Flask, render_template, request, redirect, url_for
+from flask import jsonify, flash, make_response, session as login_session
+from sqlalchemy import create_engine, desc
+from sqlalchemy.orm import sessionmaker
+from database_setup import Base, Category, Item
+from oauth2client.client import flow_from_clientsecrets, FlowExchangeError
 
 app = Flask(__name__)
 app.config['SESSION_TYPE'] = 'filesystem'
 app.config['SECRET_KEY'] = 'super secret key'
 
-CLIENT_ID = json.loads(open('client_secrets.json', 'r').read())['web']['client_id']
+file_name = 'client_secrets.json'
+CLIENT_ID = json.loads(open(file_name, 'r').read())['web']['client_id']
 
 # adding database
 engine = create_engine('sqlite:///catalog.db')
@@ -25,36 +24,57 @@ Base.metadata.bind = engine
 DBSession = sessionmaker(bind=engine)
 session = DBSession()
 
-#need template page to show categories at side every page
-#need json endpoint
+
+# JSON API endpoints
+@app.route('/categories/<string:category_name>/JSON')
+def categoryListJSON(category_name):
+    category = session.query(Category).filter_by(name=category_name).one()
+    items = session.query(Item).filter_by(category_name=category_name)
+    return jsonify(CategoryItems=[i.serialize for i in items])
+
+
+@app.route('/categories/<int:item_id>/JSON')
+def itemJSON(item_id):
+    item = session.query(Item).filter_by(id=item_id).one()
+    return jsonify(Item=item.serialize)
+
 
 @app.route('/')
 def homepage():
-    return "Need html for latest items"
+    categories = session.query(Category).all()
+    all_latest_items = session.query(Item).order_by(desc(Item.date_added))
+    latest_items = all_latest_items.limit(10).all()
+    return render_template('home.html', categories=categories,
+                           latest_items=latest_items)
+
 
 @app.route('/categories/<string:category_name>/')
 def categoryList(category_name):
+    categories = session.query(Category).all()
     category = session.query(Category).filter_by(name=category_name).one()
     items = session.query(Item).filter_by(category_name=category_name)
-    return render_template('category.html', category=category, items=items, category_name=category_name)
+    return render_template('category.html', categories=categories,
+                           category=category, items=items)
 
-@app.route('/categories/')
-def categories(category_name):
-    categories = session.query(Category).all()
-    output = ''
-    for category in categories:
-        output += category.name
-        output += '</br>' 
-    return output
 
-@app.route('/items/<string:category_name>/create/', methods = ['GET', 'POST'])
+@app.route('/categories/<string:category_name>/<int:item_id>/')
+def item(category_name, item_id):
+    category = session.query(Category).filter_by(name=category_name).one()
+    item = session.query(Item).filter_by(id=item_id).one()
+    return render_template('item.html', category=category, item=item)
+
+
+@app.route('/items/<string:category_name>/create/', methods=['GET', 'POST'])
 def createItem(category_name):
+    # directs user to login page if not logged on
+    if 'username' not in login_session:
+        return redirect('/login')
     if request.method == 'POST':
-        createdItem = Item(name = request.form['name'],
-                           description = request.form['description'],
-                           category_name = category_name)
+        createdItem = Item(name=request.form['name'],
+                           description=request.form['description'],
+                           category_name=category_name)
         session.add(createdItem)
-        session.commit
+        session.commit()
         return redirect(url_for('categoryList', category_name=category_name))
     else:
         return render_template('createitem.html', category_name=category_name)
@@ -63,6 +83,8 @@ def createItem(category_name):
 @app.route('/categories/<string:category_name>/<int:item_id>/update',
            methods=['GET', 'POST'])
 def updateItem(category_name, item_id):
+    if 'username' not in login_session:
+        return redirect('/login')
     updatedItem = session.query(Item).filter_by(id=item_id).one()
     if request.method == 'POST':
         if request.form['name']:
@@ -74,12 +96,15 @@ def updateItem(category_name, item_id):
         return redirect(url_for('categoryList', category_name=category_name))
     else:
 
-        return render_template(
-            'updateitem.html', category_name=category_name, item_id=item_id, item=updatedItem)
+        return render_template('updateitem.html', category_name=category_name,
+                               item_id=item_id, item=updatedItem)
+
 
 @app.route('/categories/<string:category_name>/<int:item_id>/delete',
            methods=['GET', 'POST'])
 def deleteItem(category_name, item_id):
+    if 'username' not in login_session:
+        return redirect('/login')
     deletingItem = session.query(Item).filter_by(id=item_id).one()
     if request.method == 'POST':
         session.delete(deletingItem)
@@ -88,91 +113,123 @@ def deleteItem(category_name, item_id):
     else:
         return render_template('deleteitem.html', item=deletingItem)
 
+
 @app.route('/login')
 def showLogin():
-    state = ''.join(random.choice(string.ascii_uppercase + string.digits) for x in xrange(32))
+    state = ''.join(random.choice(string.ascii_uppercase +
+                                  string.digits) for x in xrange(32))
     login_session['state'] = state
     return render_template('login.html', STATE=state)
 
+
+# connecting to google
 @app.route('/gconnect', methods=['POST'])
 def gconnect():
     print 'start gconnect'
-    # Only if collect one-time code if state token matches
+    # only collect one-time code if state token matches
     if request.args.get('state') != login_session['state']:
         response = make_response(json.dumps('Invalid state parameter.'), 401)
         response.headers['Content-Type'] = 'application/json'
         return response
     code = request.data
-    print 'wtf'
-    # Exchange one-time code for credentials object
+    # exchange one-time code for credentials object
     try:
-        print 'trying'
-        oauth_flow=flow_from_clientsecrets('client_secrets.json', scope='')
+        oauth_flow = flow_from_clientsecrets('client_secrets.json', scope='')
         oauth_flow.redirect_uri = 'postmessage'
-        print 'ye'
         credentials = oauth_flow.step2_exchange(code)
-        print 'no'
     except FlowExchangeError:
-        print 'butt failed'
-        response = make_response(json.dumps('Failed to upgrade the authorisation code.'), 401)
+        response_message = 'Failed to upgrade the authorisation code.'
+        response = make_response(json.dumps(response_message), 401)
         response.headers['Content-Type'] = 'application/json'
         return response
-    print 'aouch'
-    # Check for valid access token
-    access_token = login_session.get('credentials')
-    url = ('https://www.googleapis.com/oauth2/v1/tokeninfo?access_token=%s' % access_token)
+    # check for valid access token
+    access_token = credentials.access_token
+    url = ('https://www.googleapis.com/oauth2/v1/tokeninfo?access_token=%s'
+           % access_token)
     h = httplib2.Http()
-    print 'halp'
     result = json.loads(h.request(url, 'GET')[1])
     if result.get('error') is not None:
         response = make_response(json.dumps(result.get('error')), 500)
-        response.headers['Content-Type'] = 'application.json'
-    # Confirm credential id match
+        response.headers['Content-Type'] = 'application/json'
+    # confirm credential id match
     gplus_id = credentials.id_token['sub']
     if result['user_id'] != gplus_id:
         response = make_response(
             json.dumps("Token's user ID doesn't match given user ID."), 401)
         response.headers['Content-Type'] = 'application/json'
         return response
-    # Confirm client id match
-    if result['issued-to'] != CLIENT_ID:
-        response = make_response(json.dumps("Token's client ID does not match app's."), 401)
-        print "Token's client ID does not match app's."
+    # confirm client id match
+    if result['issued_to'] != CLIENT_ID:
+        response_message = "Token's client ID does not match app's."
+        response = make_response(json.dumps(response_message), 401)
         response.headers['Content-Type'] = 'application/json'
         return response
-    # Check for user already logged in
+    # check for user already logged in
     stored_credentials = login_session.get('credentials')
     stored_gplus_id = login_session.get('gplus_id')
     if stored_credentials is not None and gplus_id == stored_gplus_id:
-        print "stuch here"
-        response = make_response(json.dumps('Current user is already connected.'), 200)
+        response_message = 'Current user is already connected.'
+        response = make_response(json.dumps(response_message), 200)
         response.headers['Content-Type'] = 'application/json'
-    print 'yay'
-    # Store credentials for future use
-    login_session['credentials'] = credentials.access_token
+    # store credentials for future use
+    login_session['access_token'] = credentials.access_token
     login_session['gplus_id'] = gplus_id
 
-    # Retrieve information from Google API
-    userinfo_url = "https://www.googleapis.com/oauth2/userinfo"
+    # retrieve information from Google API
+    userinfo_url = "https://www.googleapis.com/oauth2/v1/userinfo"
     params = {'access_token': credentials.access_token, 'alt': 'json'}
     answer = requests.get(userinfo_url, params=params)
-    data = json.loads(answer.text)
+    data = answer.json()
 
     login_session['username'] = data['name']
     login_session['picture'] = data['picture']
     login_session['email'] = data['email']
 
-    #Display log in
+    # indicate logging on
     output = ''
     output += '<h1>Welcome, '
     output += login_session['username']
     output += '!</h1>'
-    output += 'img src="'
+    output += '<img src="'
     output += login_session['picture']
-    output += ' " style = "width:300px; height: 300px; border-radius: 150px;-webkit-border-radius: 150px;-moz-border-radius:150px;">'
-    flash("You are now logged in as %s"%login_session['username'])
-    print output
+    output += ' " style = "width: 300px; height: 300px;border-radius: 150px;'
+    output += '-webkit-border-radius: 150px;-moz-border-radius: 150px;"> '
+    flash("you are now logged in as %s" % login_session['username'])
     return output
+
+
+# disconnect from google
+@app.route("/gdisconnect")
+def gdisconnect():
+    # only disconnect a connected user
+    credentials = login_session.get('access_token')
+    if credentials is None:
+        response = make_response(json.dumps('No user connected.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    access_token = login_session['access_token']
+    url = 'https://accounts.google.com/o/oauth2/revoke?token=%s' % access_token
+    h = httplib2.Http()
+    result = h.request(url, 'GET')[0]
+    print 'this isssss'
+    print result
+    # upon success reset session
+    if result['status'] == '200':
+        del login_session['access_token']
+        del login_session['gplus_id']
+        del login_session['username']
+        del login_session['email']
+        del login_session['picture']
+
+        response = make_response(json.dumps('User disconnected.'), 200)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    else:
+        response_message = 'Failed to revoke token for given user.'
+        response = make_response(json.dumps(response_message), 400)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
 
 if __name__ == '__main__':
     app.debug = True
